fix(ingest): add git to Dockerfile + fix GitHub clone auth + sort all job types by active first
- Add `git` to runtime image so github ingest can clone repos (source
  cloning uses `git clone --depth=1`, requires git in container PATH)
- Switch clone auth from GIT_CONFIG_* env vars (broken in headless/no-TTY
  containers) to token-in-URL (x-access-token) with GIT_TERMINAL_PROMPT=0
- Fix list_service_jobs to sort running→pending→completed for all job kinds,
  not just crawl
- Fix pre-existing rustfmt issues in extract_ladder.rs and health_tests.rs
